breaks up gatsby-node
